Move archive package into public pkg/ directory
* Adds a README to the pkg directory, as we look into ways of sensibly organizing our codebase and exposing some ancillary packages to aspects of the community.
