+++ acceptance/acceptance_test.go
@@ -37,9 +37,9 @@ import (
 	"github.com/buildpacks/pack/acceptance/config"
 	"github.com/buildpacks/pack/acceptance/invoke"
 	pubcfg "github.com/buildpacks/pack/config"
-	"github.com/buildpacks/pack/internal/archive"
 	"github.com/buildpacks/pack/internal/cache"
 	"github.com/buildpacks/pack/internal/style"
+	"github.com/buildpacks/pack/pkg/archive"
 	h "github.com/buildpacks/pack/testhelpers"
 )
