+++ internal/blob/blob.go
@@ -9,7 +9,7 @@ import (
 	"github.com/docker/docker/pkg/ioutils"
 	"github.com/pkg/errors"
 
-	"github.com/buildpacks/pack/internal/archive"
+	"github.com/buildpacks/pack/pkg/archive"
 )
 
 type Blob interface {
