+++ internal/layer/layer.go
@@ -1,7 +1,7 @@
 package layer
 
 import (
-	"github.com/buildpacks/pack/internal/archive"
+	"github.com/buildpacks/pack/pkg/archive"
 )
 
 func CreateSingleFileTar(tarFile, path, txt string, twf archive.TarWriterFactory) error {
