+++ pkg/README.md
@@ -1,0 +1,3 @@
+`pkg/` is a collection of utility packages that are used by the Pack CLI, without being specific to its internals. Some 
+packages may get moved to `imgutil` when a project wide common API is developed, but this is a first landing spot allowing 
+us to share code with the community.
