+++ pkg/archive/archive.go
@@ -1,4 +1,5 @@
-package archive
+// Package archive defines a set of functions for reading and writing directories and files in a number of tar formats.
+package archive // import "github.com/buildpacks/pack/pkg/archive"
 
 import (
 	"archive/tar"
