+++ pkg/archive/testdata/fat-zip-to-tar.zip
